refactor(plugins): combine convert repos
before, we convert repo to repos and board seperately
now, combine them in the same task
