+++ plugins/github/github.go
@@ -84,15 +84,13 @@ func (plugin Github) Execute(options map[string]interface{}, progress chan<- flo
 			"enrichIssues":               true,
 			"enrichPullRequests":         true,
 			"convertRepos":               true,
-			"convertReposToBoard":        true,
-
-			"convertIssues":             true,
-			"convertIssueLabels":        true,
-			"convertPullRequests":       true,
-			"convertCommits":            true,
-			"convertPullRequestCommits": true,
-			"convertNotes":              true,
-			"convertUsers":              true,
+			"convertIssues":              true,
+			"convertIssueLabels":         true,
+			"convertPullRequests":        true,
+			"convertCommits":             true,
+			"convertPullRequestCommits":  true,
+			"convertNotes":               true,
+			"convertUsers":               true,
 		}
 	} else {
 		for _, task := range op.Tasks {
@@ -234,13 +232,6 @@ func (plugin Github) Execute(options map[string]interface{}, progress chan<- flo
 			return err
 		}
 	}
-	if tasksToRun["convertReposToBoard"] {
-		progress <- 0.93
-		err = tasks.ConvertReposToBoard()
-		if err != nil {
-			return err
-		}
-	}
 	if tasksToRun["convertIssues"] {
 		progress <- 0.94
 		err = tasks.ConvertIssues(repoId)
@@ -352,21 +343,21 @@ func main() {
 				"owner": owner,
 				"repo":  repo,
 				"tasks": []string{
-					"collectRepo",
+					//"collectRepo",
 					//"collectCommits",
 					//"collectCommitsStat",
-					"collectIssues",
-					"collectPullRequests",
+					//"collectIssues",
+					//"collectPullRequests",
 					//"collectIssueEvents",
 					//"collectIssueComments",
 					//"collectPullRequestReviews",
 					//"collectPullRequestCommits",
 					//"collectPullRequestComments",
-					"enrichIssues",
-					"enrichPullRequests",
+					//"enrichIssues",
+					//"enrichPullRequests",
 					"convertRepos",
-					"convertIssues",
-					"convertPullRequests",
+					//"convertIssues",
+					//"convertPullRequests",
 					//"convertCommits",
 					//"convertPullRequestCommits",
 					//"convertNotes",
