+++ plugins/github/tasks/github_repo_converter.go
@@ -5,6 +5,7 @@ import (
 	"github.com/merico-dev/lake/models/domainlayer"
 	"github.com/merico-dev/lake/models/domainlayer/code"
 	"github.com/merico-dev/lake/models/domainlayer/didgen"
+	"github.com/merico-dev/lake/models/domainlayer/ticket"
 	githubModels "github.com/merico-dev/lake/plugins/github/models"
 	"gorm.io/gorm/clause"
 )
@@ -15,15 +16,23 @@ func ConvertRepos() error {
 	if err != nil {
 		return err
 	}
-	userIdGen := didgen.NewDomainIdGenerator(&githubModels.GithubUser{})
+	domainUserIdGenerator := didgen.NewDomainIdGenerator(&githubModels.GithubUser{})
+	domainRepoIdGenerator := didgen.NewDomainIdGenerator(&githubModels.GithubRepo{})
+
 	for _, repository := range githubRepositorys {
 		domainRepository := convertToRepositoryModel(&repository)
-		domainRepository.OwnerId = userIdGen.Generate(repository.OwnerId)
+		domainRepository.OwnerId = domainUserIdGenerator.Generate(repository.OwnerId)
 		err := lakeModels.Db.Clauses(clause.OnConflict{UpdateAll: true}).Create(domainRepository).Error
 		if err != nil {
 			return err
 		}
+		domainBoard := convertToBoardModel(&repository, domainRepoIdGenerator)
+		err = lakeModels.Db.Clauses(clause.OnConflict{UpdateAll: true}).Create(domainBoard).Error
+		if err != nil {
+			return err
+		}
 	}
+
 	return nil
 }
 func convertToRepositoryModel(repository *githubModels.GithubRepo) *code.Repo {
@@ -40,4 +49,17 @@ func convertToRepositoryModel(repository *githubModels.GithubRepo) *code.Repo {
 		UpdatedDate: repository.UpdatedDate,
 	}
 	return domainRepository
+}
+func convertToBoardModel(repository *githubModels.GithubRepo, domainIdGenerator *didgen.DomainIdGenerator) *ticket.Board {
+	domainBoard := &ticket.Board{
+		DomainEntity: domainlayer.DomainEntity{
+			Id: domainIdGenerator.Generate(repository.GithubId),
+		},
+		Name:        repository.Name,
+		Url:         repository.HTMLUrl,
+		Description: repository.Description,
+
+		CreatedDate: &repository.CreatedDate,
+	}
+	return domainBoard
 }
